Refactor ray functionality
