+++ alibi/explainers/anchor_tabular_distributed.py
@@ -12,16 +12,93 @@ from alibi.api.interfaces import Explainer, Explanation, FitMixin
 from alibi.exceptions import (AlibiPredictorCallException,
                               AlibiPredictorReturnTypeError)
 from alibi.utils.discretizer import Discretizer
-from alibi.utils.distributed import RAY_INSTALLED
 from alibi.utils.mapping import ohe_to_ord, ord_to_ohe
 from alibi.utils.wrappers import ArgmaxTransformer
 
-from .anchor_base import AnchorBaseBeam, DistributedAnchorBaseBeam
+from .anchor_base import AnchorBaseBeam
 from .anchor_explanation import AnchorExplanation
 from .anchor_tabular import AnchorTabular, TabularSampler
-
+from alibi.utils.distributed import ActorPool
+from functools import partial
 
 import ray
+
+
+class DistributedAnchorBaseBeam(AnchorBaseBeam):
+    ray = ray  #: `ray` module.
+
+    def __init__(self, samplers: List[Callable], **kwargs) -> None:
+
+        super().__init__(samplers)
+        self.chunksize = kwargs.get('chunksize', 1)
+        self.sample_fcn = lambda actor, anchor, n_samples, compute_labels=True: \
+            actor.__call__.remote(anchor,
+                                  n_samples,
+                                  compute_labels=compute_labels)
+        self.pool = ActorPool(samplers)
+        self.samplers = samplers
+
+    def _get_coverage_samples(self, coverage_samples: int,  # type: ignore[override]
+                              samplers: List[Callable]) -> np.ndarray:
+        """
+        Sends a request for a coverage set to process running sampling tasks.
+
+        Parameters
+        ----------
+        coverage_samples, samplers
+            See :py:meth:`alibi.explainers.anchor_base.AnchorBaseBeam._get_coverage_samples` implementation.
+
+        Returns
+        -------
+        See :py:meth:`alibi.explainers.anchor_base.AnchorBaseBeam._get_coverage_samples` implementation.
+        """
+
+        [coverage_data] = DistributedAnchorBaseBeam.ray.get(
+            self.sample_fcn(samplers[0], (0, ()), coverage_samples, compute_labels=False)
+        )
+
+        return coverage_data
+
+    def draw_samples(self, anchors: list, batch_size: int) -> Tuple[np.ndarray, np.ndarray]:  # type: ignore[override]
+        """
+        Distributes sampling requests among processes running sampling tasks.
+
+        Parameters
+        ----------
+        anchors, batch_size
+            See :py:meth:`alibi.explainers.anchor_base.AnchorBaseBeam.draw_samples` implementation.
+
+        Returns
+        -------
+        See :py:meth:`alibi.explainers.anchor_base.AnchorBaseBeam.draw_samples` implementation.
+        """
+
+        # partial anchors not generated by propose_anchors are not in the order dictionary
+        for anchor in anchors:
+            if anchor not in self.state['t_order']:
+                self.state['t_order'][anchor] = list(anchor)
+
+        pos, total = np.zeros((len(anchors),)), np.zeros((len(anchors),))
+        order_map = [(i, tuple(self.state['t_order'][anchor])) for i, anchor in enumerate(anchors)]
+        samples_iter = self.pool.map_unordered(
+            partial(self.sample_fcn, n_samples=batch_size),
+            order_map,
+            self.chunksize,
+        )
+        for samples_batch in samples_iter:
+            for samples in samples_batch:
+                covered_true, covered_false, labels, *additionals, anchor_idx = samples
+                positives, n_samples = self.update_state(
+                    covered_true,
+                    covered_false,
+                    labels,
+                    additionals,
+                    anchors[anchor_idx],
+                )
+                # return statistics in the same order as the requests
+                pos[anchor_idx], total[anchor_idx] = positives, n_samples
+
+        return pos, total
 
 
 class RemoteSampler:
